main/requirements.txt
pyserial
numpy
pynmea2
tinydb
